Module aws_sdk_ram::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptResourceShareInvitation

Fluent builder constructing a request to AcceptResourceShareInvitation.

AssociateResourceShare

Fluent builder constructing a request to AssociateResourceShare.

AssociateResourceSharePermission

Fluent builder constructing a request to AssociateResourceSharePermission.

CreateResourceShare

Fluent builder constructing a request to CreateResourceShare.

DeleteResourceShare

Fluent builder constructing a request to DeleteResourceShare.

DisassociateResourceShare

Fluent builder constructing a request to DisassociateResourceShare.

DisassociateResourceSharePermission

Fluent builder constructing a request to DisassociateResourceSharePermission.

EnableSharingWithAwsOrganization

Fluent builder constructing a request to EnableSharingWithAwsOrganization.

GetPermission

Fluent builder constructing a request to GetPermission.

GetResourcePolicies

Fluent builder constructing a request to GetResourcePolicies.

GetResourceShareAssociations

Fluent builder constructing a request to GetResourceShareAssociations.

GetResourceShareInvitations

Fluent builder constructing a request to GetResourceShareInvitations.

GetResourceShares

Fluent builder constructing a request to GetResourceShares.

ListPendingInvitationResources

Fluent builder constructing a request to ListPendingInvitationResources.

ListPermissions

Fluent builder constructing a request to ListPermissions.

ListPrincipals

Fluent builder constructing a request to ListPrincipals.

ListResourceSharePermissions

Fluent builder constructing a request to ListResourceSharePermissions.

ListResourceTypes

Fluent builder constructing a request to ListResourceTypes.

ListResources

Fluent builder constructing a request to ListResources.

PromoteResourceShareCreatedFromPolicy

Fluent builder constructing a request to PromoteResourceShareCreatedFromPolicy.

RejectResourceShareInvitation

Fluent builder constructing a request to RejectResourceShareInvitation.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateResourceShare

Fluent builder constructing a request to UpdateResourceShare.