Module aws_sdk_ram::error

Modules

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

invalid_client_token_exception

See InvalidClientTokenException

invalid_max_results_exception

See InvalidMaxResultsException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_resource_type_exception

See InvalidResourceTypeException

invalid_state_transition_exception

See InvalidStateTransitionException

malformed_arn_exception

See MalformedArnException

missing_required_parameter_exception

See MissingRequiredParameterException

operation_not_permitted_exception

See OperationNotPermittedException

resource_arn_not_found_exception

See ResourceArnNotFoundException

resource_share_invitation_already_accepted_exception

See ResourceShareInvitationAlreadyAcceptedException

resource_share_invitation_already_rejected_exception

See ResourceShareInvitationAlreadyRejectedException

resource_share_invitation_arn_not_found_exception

See ResourceShareInvitationArnNotFoundException

resource_share_invitation_expired_exception

See ResourceShareInvitationExpiredException

resource_share_limit_exceeded_exception

See ResourceShareLimitExceededException

server_internal_exception

See ServerInternalException

service_unavailable_exception

See ServiceUnavailableException

tag_limit_exceeded_exception

See TagLimitExceededException

tag_policy_violation_exception

See TagPolicyViolationException

unknown_resource_exception

See UnknownResourceException

Structs

AcceptResourceShareInvitationError

AssociateResourceShareError

AssociateResourceSharePermissionError

CreateResourceShareError

DeleteResourceShareError

DisassociateResourceShareError

DisassociateResourceSharePermissionError

EnableSharingWithAwsOrganizationError

GetPermissionError

GetResourcePoliciesError

GetResourceShareAssociationsError

GetResourceShareInvitationsError

GetResourceSharesError

IdempotentParameterMismatchException

A client token input parameter was reused with an operation, but at least one of the

InvalidClientTokenException

A client token is not valid.

InvalidMaxResultsException

The specified value for MaxResults is not valid.

InvalidNextTokenException

The specified value for NextToken is not valid.

InvalidParameterException

A parameter is not valid.

InvalidResourceTypeException

The specified resource type is not valid.

InvalidStateTransitionException

The requested state transition is not valid.

ListPendingInvitationResourcesError

ListPermissionsError

ListPrincipalsError

ListResourceSharePermissionsError

ListResourceTypesError

ListResourcesError

MalformedArnException

The format of an Amazon Resource Name (ARN) is not valid.

MissingRequiredParameterException

A required input parameter is missing.

OperationNotPermittedException

The requested operation is not permitted.

PromoteResourceShareCreatedFromPolicyError

RejectResourceShareInvitationError

ResourceArnNotFoundException

An Amazon Resource Name (ARN) was not found.

ResourceShareInvitationAlreadyAcceptedException

The invitation was already accepted.

ResourceShareInvitationAlreadyRejectedException

The invitation was already rejected.

ResourceShareInvitationArnNotFoundException

The Amazon Resource Name (ARN) for an invitation was not found.

ResourceShareInvitationExpiredException

The invitation is expired.

ResourceShareLimitExceededException

The requested resource share exceeds the limit for your account.

ServerInternalException

The service could not respond to the request due to an internal problem.

ServiceUnavailableException

The service is not available.

TagLimitExceededException

The requested tags exceed the limit for your account.

TagPolicyViolationException

The specified tag is a reserved word and cannot be used.

TagResourceError

UnknownResourceException

A specified resource was not found.

UntagResourceError

UpdateResourceShareError

Enums

AcceptResourceShareInvitationErrorKind

AssociateResourceShareErrorKind

AssociateResourceSharePermissionErrorKind

CreateResourceShareErrorKind

DeleteResourceShareErrorKind

DisassociateResourceShareErrorKind

DisassociateResourceSharePermissionErrorKind

EnableSharingWithAwsOrganizationErrorKind

GetPermissionErrorKind

GetResourcePoliciesErrorKind

GetResourceShareAssociationsErrorKind

GetResourceShareInvitationsErrorKind

GetResourceSharesErrorKind

ListPendingInvitationResourcesErrorKind

ListPermissionsErrorKind

ListPrincipalsErrorKind

ListResourceSharePermissionsErrorKind

ListResourceTypesErrorKind

ListResourcesErrorKind

PromoteResourceShareCreatedFromPolicyErrorKind

RejectResourceShareInvitationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateResourceShareErrorKind