Module aws_sdk_ram::model

Modules

principal

See Principal

resource

See Resource

resource_share

See ResourceShare

resource_share_association

See ResourceShareAssociation

resource_share_invitation

See ResourceShareInvitation

resource_share_permission_detail

See ResourceSharePermissionDetail

resource_share_permission_summary

See ResourceSharePermissionSummary

service_name_and_resource_type

See ServiceNameAndResourceType

tag

See Tag

tag_filter

See TagFilter

Structs

Principal

Describes a principal for use with Resource Access Manager.

Resource

Describes a resource associated with a resource share.

ResourceShare

Describes a resource share.

ResourceShareAssociation

Describes an association with a resource share.

ResourceShareInvitation

Describes an invitation to join a resource share.

ResourceSharePermissionDetail

Information about an RAM permission.

ResourceSharePermissionSummary

Information about a permission that is associated with a resource share.

ServiceNameAndResourceType

Information about the shareable resource types and the Amazon Web Services services to which they

Tag

Information about a tag.

TagFilter

Used to filter information based on tags.

Enums

ResourceOwner

NOTE: ResourceOwner::Self has been renamed to ::SelfValue.

ResourceShareAssociationStatus

ResourceShareAssociationType

ResourceShareFeatureSet

ResourceShareInvitationStatus

ResourceShareStatus

ResourceStatus