Module aws_sdk_ram::operation

Structs

AcceptResourceShareInvitation

Accepts an invitation to a resource share from another Amazon Web Services account.

AssociateResourceShare

Associates the specified resource share with the specified principals and resources.

AssociateResourceSharePermission

Associates a permission with a resource share.

CreateResourceShare

Creates a resource share. You must provide a list of the Amazon Resource Names (ARNs) for the

DeleteResourceShare

Deletes the specified resource share.

DisassociateResourceShare

Disassociates the specified principals or resources from the specified resource share.

DisassociateResourceSharePermission

Disassociates an RAM permission from a resource share.

EnableSharingWithAwsOrganization

Enables resource sharing within your organization in Organizations.

GetPermission

Gets the contents of an RAM permission in JSON format.

GetResourcePolicies

Gets the policies for the specified resources that you own and have shared.

GetResourceShareAssociations

Gets the resources or principals for the resource shares that you own.

GetResourceShareInvitations

Gets the invitations that you have received for resource shares.

GetResourceShares

Gets the resource shares that you own or the resource shares that are shared with you.

ListPendingInvitationResources

Lists the resources in a resource share that is shared with you but that the invitation is still

ListPermissions

Lists the RAM permissions.

ListPrincipals

Lists the principals that you have shared resources with or that have shared resources

ListResourceSharePermissions

Lists the RAM permissions that are associated with a resource share.

ListResourceTypes

Lists the shareable resource types supported by RAM.

ListResources

Lists the resources that you added to a resource shares or the resources that are shared with

PromoteResourceShareCreatedFromPolicy

Resource shares that were created by attaching a policy to a resource are visible only

RejectResourceShareInvitation

Rejects an invitation to a resource share from another Amazon Web Services account.

TagResource

Adds the specified tags to the specified resource share that you own.

UntagResource

Removes the specified tags from the specified resource share that you own.

UpdateResourceShare

Updates the specified resource share that you own.